backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clerk-backend-api",
    "fastapi>=0.121.1",
    "openai>=2.7.2",
    "google-generativeai>=0.5.4",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "svix>=1.81.0",
]
